# [MG-4] Google 소셜 로그인 구현 완료 및 메일에 대한 임베딩 검색 구현