--- a/main.py
+++ b/main.py
@@ -1,18 +1,20 @@
-from fastapi import Depends, FastAPI
-from sqlalchemy.orm import Session
-from database import SessionLocal
-from models import User
+from fastapi import FastAPI
+from starlette.middleware.sessions import SessionMiddleware
+from dotenv import load_dotenv
+from controllers import routers
+from logger import get_logger
+import os
 
+# 환경변수 로드
+load_dotenv()
+
+# FastAPI 앱 초기화
 app = FastAPI()
+logger = get_logger("main")
 
-# DB 세션 의존성
-def get_db():
-    db = SessionLocal()
-    try:
-        yield db
-    finally:
-        db.close()
+# 세션 미들웨어 추가
+app.add_middleware(SessionMiddleware, secret_key=os.getenv('JWT_SECRET_KEY'))
 
-@app.get("/users")
-def read_users(db: Session = Depends(get_db)):
-    return db.query(User).all()
+# 라우터 등록
+for router in routers:
+    app.include_router(router)