--- a/requirements.txt
+++ b/requirements.txt
@@ -10,3 +10,5 @@ starlette==0.46.2
 typing-inspection==0.4.0
 typing_extensions==4.13.2
 uvicorn==0.34.2
+sqlalchemy
+pymysql