## 📝 PR 타입

- [x] 기능 구현
- [ ] 기능 수정
- [ ] 버그 수정
- [ ] 리팩토링
- [ ] 인프라, 의존성, 환경 변수, 빌드 관련 코드 업데이트

## 🚀 변경 사항

1. 프로젝트를 위한 파일을 재구성했습니다.
2. 로그인 시, 소셜 로그인 기능을 구현하였습니다.
3. 메일 전체를 읽어와, 벡터 임베딩 후 벡터 DB 검색을 구현했습니다.

### 프로젝트 구조 변경

Spring에서 사용했던 MVC 구조와 동일하게 변경하였습니다.
전체적인 흐름은 controller -> service -> model로 흘러가며, 각 Layer에 대한 폴더 및 파일 구조를 생성했습니다. 
코드 읽어보시면서 흐름 봐보시면 좋을 것 같습니다!

### 로그인 구현 및 사용 방법

 구글 로그인 Oauth 2.0 방식으로 구현하였습니다.
앱의 테스트 계정에 이메일을 추가하시면, 로그인이 가능합니다!

로그인 방법은 다음과 같습니다. 

1. 브라우저에서 /google URI로 요청 -> 구글 로그인 화면이 뜨면, 로그인 진행
2. 로그인 진행 완료되면, user 테이블에 유저 정보가 추가되고, response로 유저 정보를 보여주게 됩니다. -> json이 보인다면 로그인 성공!

로그인이 완료되면, python 서버에서 발급해준 고유 token이 세션에 들어가게 되고, gmail 요청마다 그 세션 토큰을 사용해 GMAIL API 호출 및 처리를 하게 됩니다. 즉, 세션 토큰이 없으면 gmail 관련 api 호출이 불가능함(Postman 사용 불가)

![image](https://github.com/user-attachments/assets/f83caf7a-5b8a-43e8-8d8d-69a9e4a1a8fd)

### 벡터 임베딩 구현

어제 논의했던 내용 중, 키워드 추출 방식 및 삭제할 메일을 어떻게 추천해줄 것인가를 고민해보았습니다.
민경님이 말씀해주신 일반적으로 스팸에 대한 키워드들이 들어있는 메일들을 보여주고, 사용자들이 선택해서 메일을 걸러내는 방식을 조금 더 확장해봤습니다.

우선 전 키워드를 고정하는 방식은 좋아보였고, 완전일치보다는 다른 검색 방식(연관 검색)을 구현해보았습니다.

제가 구현한 Flow는 이렇습니다.
1. 메일 전체를 받아온다.
2. 각 메일의 제목 + 내용 일부를 Vector DB에 임베딩
3. 키워드(광고) 검색해, 연관 메일을 검색

메일이 총 16000개라, 한번에 처리하기에는 메모리와 임베딩 속도가 부담스러워서 Redis + Celery로 분산 작업을 통해 구현했습니다.

### 1. 메일 전체 받아오기

구글 메일 API는 여러개의 메일에 대한 상세정보를 한번에 가져오는 API를 지원하지 않습니다.
한번에 여러개의 메일에 대한 정보를 받아오려면 ID 목록만 받아올 수 있었습니다.
그래서 한번에 메일 ID들을 받아오는 API를 통해 500개씩 ID 받아왔고, 각각 하나하나 호출해 내용을 가져오는 식으로 구현했습니다.

제 메일함은 총 16000개의 메일이 들어있어, 500개씩 32번 반복하고, 1000개로 쪼개 Celery로 보내주게 됩니다.

 큐에 모두 넣고 나면, 클라이언트에게 완료 Response를 보내주게 됩니다.

### 2. 각 메일의 제목 + 내용 일부를 Vector DB에 임베딩

완료 Response 이후부터, Celery에서 처리를 시작합니다.
Celery에서 1000개씩 처리할 때 다음 과정을 거치게 됩니다.

1. ID를 통해 메일 내용 호출
=> 기존에 1개씩 불러오면, 요청 -> 응답 NetworkLatency로 1개당 2~3초정도 걸렸습니다.(1000개면 2000초...)
=> http 요청을 한번에 batch로 날리고, 받아온 걸 한번에 모으는 방식으로 변경했습니다.(2000초 => 최대 10초 정도로 감소)

![image](https://github.com/user-attachments/assets/965b2250-61e0-4a7b-add7-23e0809c787a)


2. 1000개를 한번에 embedding 시키기
모든 텍스트를 한번에 임베딩해, chroma DB에 저장하게 됩니다.

```text
[2025-04-30 15:39:43,987: INFO/MainProcess] 임베딩 시작: 981개 텍스트
Batches:   0%|          | 0/31 [00:00<?, ?it/s]
Batches:   3%|3         | 1/31 [00:00<00:27,  1.07it/s]
Batches:   6%|6         | 2/31 [00:01<00:21,  1.38it/s]
Batches:  10%|9         | 3/31 [00:02<00:18,  1.49it/s]
Batches:  13%|#2        | 4/31 [00:02<00:17,  1.56it/s]
Batches:  16%|#6        | 5/31 [00:03<00:16,  1.61it/s]
Batches:  19%|#9        | 6/31 [00:03<00:15,  1.63it/s]
Batches:  23%|##2       | 7/31 [00:04<00:14,  1.65it/s]
Batches:  26%|##5       | 8/31 [00:05<00:13,  1.67it/s]
Batches:  29%|##9       | 9/31 [00:05<00:13,  1.67it/s]
Batches:  32%|###2      | 10/31 [00:06<00:12,  1.67it/s]
Batches:  35%|###5      | 11/31 [00:06<00:11,  1.68it/s]
Batches:  39%|###8      | 12/31 [00:07<00:11,  1.68it/s]
Batches:  42%|####1     | 13/31 [00:08<00:10,  1.68it/s]
Batches:  45%|####5     | 14/31 [00:08<00:10,  1.68it/s]
Batches:  48%|####8     | 15/31 [00:09<00:09,  1.68it/s]
Batches:  52%|#####1    | 16/31 [00:09<00:08,  1.68it/s]
Batches:  55%|#####4    | 17/31 [00:10<00:08,  1.69it/s]
Batches:  58%|#####8    | 18/31 [00:11<00:07,  1.69it/s]
Batches:  61%|######1   | 19/31 [00:11<00:07,  1.70it/s]
Batches:  65%|######4   | 20/31 [00:12<00:06,  1.70it/s]
Batches:  68%|######7   | 21/31 [00:12<00:05,  1.69it/s]
Batches:  71%|#######   | 22/31 [00:13<00:05,  1.69it/s]
Batches:  74%|#######4  | 23/31 [00:13<00:04,  1.69it/s]
Batches:  77%|#######7  | 24/31 [00:14<00:04,  1.69it/s]
Batches:  81%|########  | 25/31 [00:15<00:03,  1.69it/s]
Batches:  84%|########3 | 26/31 [00:15<00:02,  1.69it/s]
Batches:  87%|########7 | 27/31 [00:16<00:02,  1.69it/s]
Batches:  90%|######### | 28/31 [00:16<00:01,  1.69it/s]
Batches:  94%|#########3| 29/31 [00:17<00:01,  1.69it/s]
Batches:  97%|#########6| 30/31 [00:18<00:00,  1.70it/s]
Batches: 100%|##########| 31/31 [00:18<00:00,  1.85it/s]
Batches: 100%|##########| 31/31 [00:18<00:00,  1.67it/s]
[2025-04-30 15:40:02,557: INFO/MainProcess] 임베딩 완료: 981개 벡터
```
이런식으로 로그를 찍어보니, 배치 하나당 20초, 17개 배치로 가정하면 340초, 즉 5분~6분사이에 16000개 메일 처리가 가능합니다.
시간 측정해보니 6분 10초 정도 걸렸습니다!
(M1 Macbook 기준)

이정도 시간이면 SQS + Lambda 최대 컴퓨팅 시간인 15분 내에 가능할 것 같은데, Lambda CPU가 잘 따라와줄지는 모르겠네요..!!

로컬 기준으로 1000개 메일 정도는 20초~30초면 모든 임베딩 프로세스가 끝날 것 같습니다!

![image](https://github.com/user-attachments/assets/4d8e3d9e-c754-4f44-8b78-5d7eb59487c3)

### 3. 키워드(광고) 검색해, 연관 메일을 검색 (여기서부터는 Postman 사용 가능!)

이렇게 모두 임베딩 한 이후에, 검색을 돌려보았더니, 유사도가 상당히..유사한 부분도 있고 아닌 부분도 있었습니다.
광고가 아닌데도 뜨는 경우도 많았어서, 이런 경우를 필터링하려면 chroma DB 칼럼을 대주제를 넣고(ex) 광고), 대주제에 해당되는 것들 중에서 연관 검색을 하는 방식이 좋을 것 같습니다!

또한 메일과 연관된 메일도 찾을 수 있다고 하니, 사용자가 특정 메일을 골랐을 때 연관 메일도 추천하는 방식으로 구성하는 것도 좋아보입니다.

## 💡 To Reviewer

.env 인수인계가 필요하니, env 파일은 따로 공유하겠습니다.
또한 Vector DB의 정확도가 떨어져, 회의에 언급하셨던 내용인 카테고리 별 검색 + 임베딩 연관성 검색 두가지 방식으로 검색하는 것이 좋아보입니다.

한번 회의해보시고 방법 괜찮으시면, 처리를 개선하는 방향으로 개발하는 것도 좋을 것 같습니다!

## 🧪 테스트 결과

1. 배치 완료했을 때 Response
![image](https://github.com/user-attachments/assets/044a09fb-c148-4c02-9ebf-4836166a9b4a)

2. 검색 결과 Response

**JIRA로 검색** 
![image](https://github.com/user-attachments/assets/0ea0ee2e-7e9c-4157-9e17-fc943d078969)

**금융으로 검색**
![image](https://github.com/user-attachments/assets/3ee6871a-0cd5-4043-84f7-3a94112cff51)

사진들을 보시면 정확도가 상당히 낮은 것을 볼 수 있습니다!

## 개선 여지

코드를 작성해보면서 생각해본 아이디어입니다.

### 임베딩 싱글 쓰레드
- 현재 celery가 임베딩 엔진을 돌리기 위해서는 싱글 쓰레드를 사용해야 해서, 다소 느린 부분이 있습니다(m1 코어 한개만 사용중)
=> 멀티 쓰레드로 임베딩이 가능하다면, 훨씬 빠른 결과를 볼 수 있을 것 같습니다.

### 대주제별 분산 큐 구성

- 분류하는 Celery 큐 하나,  대주제별로 Celery 큐 다수를 구성한다고 가정하면,
=> 분류하는 Celery 큐 -> 주제별 Celery 큐로 분배하여, 단계적으로 처리하면 조금 더 빨라지지 않을까 이론적으로 생각해봅니다..!!

## ✅ 반영 브랜치
main

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- **New Features**
  - Introduced Google OAuth2 authentication, enabling users to log in with their Google accounts.
  - Added endpoints for retrieving Gmail messages, embedding email content, and searching emails by semantic similarity.
  - Implemented user management endpoints to view registered users.
  - Provided a structured GitHub pull request template for contributors.

- **Bug Fixes**
  - Improved error handling and logging across authentication, email processing, and user endpoints.

- **Chores**
  - Updated `.gitignore` to exclude Redis and Chroma storage files.
  - Added new dependencies for database support and asynchronous task processing.
  - Centralized configuration and logging setup for easier environment management.

- **Refactor**
  - Modularized application structure by separating authentication, Gmail, and user logic into dedicated controllers.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->